pyproject.toml
[tool.poetry]
name = "innotter"
version = "0.1.0"
description = ""
authors = ["Artyom Grishan <artemgris66@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
celery = "^5.2.7"
flake8 = "^4.0.1"
isort = "^5.10.1"
psycopg2 = "^2.9.3"
pyflakes = "^2.4.0"
pytest = "^7.1.2"
redis = "^4.3.3"
sqlparse = "^0.4.2"
python-decouple = "^3.6"
PyJWT = "^2.4.0"
django-extensions = "^3.1.5"
drf-extensions = "^0.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
